// DXC_Frontend/src/api/models/ocopProductDto.ts
/**
 * Generated by orval v7.11.2 🍺
 * Do not edit manually.
 * DXC_Core.API
 * OpenAPI spec version: 1.0
 */
import type { OcopProductCategoryDto } from './ocopProductCategoryDto'
import type { OcopEnterpriseDto } from './ocopEnterpriseDto'
import type { OcopProductImageDto } from './ocopProductImageDto'

export interface OcopProductDto {
  publicId?: string
  /** @nullable */
  name: string | null
  /** @nullable */
  description?: string | null
  /** @nullable */
  referencePrice?: number | null
  /** @nullable */
  promotionalPrice?: number | null
  /** @nullable */
  contactPhone?: string | null
  /** @nullable */
  contactAddress?: string | null
  /** @nullable */
  latitude?: number | null
  /** @nullable */
  longitude?: number | null
  isActive?: boolean
  createdAt?: string
  /** @nullable */
  updatedAt?: string | null
  category?: OcopProductCategoryDto
  enterprise?: OcopEnterpriseDto
  /** @nullable */
  images?: OcopProductImageDto[] | null
}